Fix formatting issues and update dependencies

+++ lib/data/adapters/theme_type_adapter.dart
@@ -8,7 +8,7 @@ class ThemeTypeAdapter extends TypeAdapter<ThemeType> {
 
   @override
   ThemeType read(BinaryReader reader) {
-    return ThemeType.values[reader.read()];
+    return ThemeType.values[reader.read() as int];
   }
 
   @override
